requirements.txt
fastapi==0.87.0
requests==2.25.1
uvicorn[standard]==0.13.4
python-dotenv==0.17.1
python-multipart==0.0.5
python-jose[cryptography]==3.2.0
passlib[bcrypt]==1.7.4
jinja2==3.0.1
cryptography==3.4.8
authlib==1.1.0